Reorg twizzler-object
